refactor/rename component
