+++ src/features/tutorApplication/index.module.scss
@@ -24,7 +24,7 @@
   }
 }
 
-.questionnaire {
+.application {
   display: flex;
   flex-direction: column;
   gap: 40px;
